Move CorsUtils into separate file
